Fix bug in parameter parsing for Yao

+++ de.tudarmstadt.maki.tc.cbctc.algorithms/src/de/tudarmstadt/maki/tc/cbctc/algorithms/facade/EMoflonFacade.java
@@ -173,11 +173,12 @@ public class EMoflonFacade extends TopologyControlFacade_ImplBase
 
       if (this.algorithm instanceof YaoGraphAlgorithm)
       {
-         final Integer coneCount = parameters.getInt(UnderlayTopologyControlAlgorithms.YAO_PARAM_CONE_COUNT);
+         //TODO@rkluge: Fix getInt
+         final Double coneCount = parameters.getDouble(UnderlayTopologyControlAlgorithms.YAO_PARAM_CONE_COUNT);
          if (coneCount == null)
             throw new IllegalArgumentException(
                   String.format("Missing mandatory parameter '%s' for %s", UnderlayTopologyControlAlgorithms.YAO_PARAM_CONE_COUNT, this.getConfiguredAlgorithm()));
-         YaoGraphAlgorithm.class.cast(this.algorithm).setConeCount(coneCount);
+         YaoGraphAlgorithm.class.cast(this.algorithm).setConeCount(coneCount.intValue());
       }
 
       this.algorithm.initializeConstraints();
